requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20

pdfplumber==0.11.4
pydantic==2.10.3

neo4j==5.27.0

langchain==0.3.13
langchain-groq==0.2.1
langchain-core==0.3.28
langgraph==0.2.59

python-dotenv==1.0.1

httpx==0.28.1
